@@ -9,55 +9,68 @@
   },
   "author": "",
   "license": "ISC",
+  "browserslist": [
+    "last 2 version"
+  ],
   "devDependencies": {
-    "@types/googlemaps": "^3.30.0",
-    "@types/node": "^8.0.53",
-    "@types/react": "^16.0.25",
-    "flatpickr": "^4.1.2",
+    "@types/datatables.net": "^1.10.8",
+    "@types/googlemaps": "^3.30.5",
+    "@types/jquery": "^3.3.0",
+    "@types/moment-duration-format": "^2.2.0",
+    "@types/node": "^9.4.0",
+    "@types/react": "^16.0.36",
     "@types/react-dom": "^16.0.3",
+    "@types/select2": "^4.0.41",
     "assets-webpack-plugin": "^3.5.1",
-    "awesome-typescript-loader": "^3.3.0",
+    "autoprefixer": "^7.2.5",
+    "awesome-typescript-loader": "^3.4.1",
     "babel-core": "^6.26.0",
     "babel-loader": "^7.1.2",
-    "babel-polyfill": "^6.26.0",
     "babel-preset-env": "^1.6.1",
-    "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "copy-webpack-plugin": "^4.2.1",
-    "css-loader": "^0.28.7",
+    "bootstrap": "^4.0.0",
+    "clean-webpack-plugin": "^0.1.18",
+    "copy-webpack-plugin": "^4.3.1",
+    "css-loader": "^0.28.9",
+    "css-split-webpack-plugin": "^0.2.5",
     "cssnano": "^3.10.0",
-    "es6-promise": "^4.1.1",
-    "expose-loader": "^0.7.3",
+    "es6-promise": "^4.2.4",
+    "expose-loader": "^0.7.4",
     "extract-text-webpack-plugin": "^3.0.2",
-    "file-loader": "^1.1.5",
+    "file-loader": "^1.1.6",
+    "flatpickr": "^4.2.4",
+    "hard-source-webpack-plugin": "^0.5.16",
+    "install": "^0.10.4",
     "isomorphic-fetch": "^2.2.1",
-    "moment": "^2.19.2",
-    "node-sass": "^4.6.1",
+    "less": "^2.7.3",
+    "less-loader": "^4.0.5",
+    "moment": "^2.20.1",
+    "node-sass": "^4.7.2",
     "normalizr": "^3.2.4",
+    "npm": "^5.6.0",
     "optimize-css-assets-webpack-plugin": "^3.2.0",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
+    "postcss-cssnext": "^3.1.0",
+    "postcss-import": "^11.0.0",
+    "postcss-loader": "^2.0.10",
+    "pug": "^2.0.0-rc.4",
+    "pug-html-loader": "^1.1.5",
+    "react": "^16.2.0",
+    "react-dom": "^16.2.0",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
     "redux-devtools": "^3.4.1",
     "redux-thunk": "^2.2.0",
     "sass-loader": "^6.0.6",
-    "socketcluster-client": "^9.0.0",
+    "socketcluster-client": "^9.0.3",
     "source-map-loader": "^0.2.3",
-    "style-loader": "^0.19.0",
-    "typescript": "^2.6.1",
-    "url-join": "^2.0.2",
+    "style-loader": "^0.20.1",
+    "typescript": "^2.7.1",
+    "url-join": "^3.0.0",
     "url-loader": "^0.6.2",
-    "webpack": "^3.8.1",
+    "webpack": "^3.10.0",
     "webpack-merge": "^4.1.1"
   },
   "dependencies": {
-    "@types/bootstrap": "^3.3.36",
-    "@types/datatables.net": "^1.10.7",
-    "@types/jquery": "^3.2.16",
-    "@types/moment-duration-format": "^1.3.8",
-    "install": "^0.10.2",
-    "npm": "^5.5.1"
+    "@types/bootstrap": "^3.3.37"
   }
 }
