@@ -2,7 +2,7 @@ let path = require('path');
 
 module.exports = function () {
     return {
-        public: 'https://application-dashboard.teraception.com/',
+        public: '/',
         contentOutput: path.join(__dirname,'public'),
         font: 'fonts',
         images: 'images',
