@@ -2,6 +2,6 @@ import TimeHelper from '../Utility/time';
 import '../startup';
 require('../../sass/common.scss');
 import '../FormSubmitters/FormMapper';
-require('../globals');
+import '../globals';
 import CrudHelper from "../Crud/CrudHelper";
 import BaseModalHandler from "../ModalHandler/BaseModalHandler";
