@@ -1,2 +1,2 @@
 require('../../../sass/moneymaker.scss');
-require('../../specific/moneymaker/transaction');
+import '../../specific/moneymaker/transaction';
