@@ -1,12 +1,13 @@
-var path = require('path');
-var webpack = require('webpack');
-var urljoin = require('url-join');
-var CopyWebpackPlugin = require('copy-webpack-plugin');
+const path = require('path');
+const webpack = require('webpack');
+const urljoin = require('url-join');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
 const ExtractTextPlugin = require("extract-text-webpack-plugin");
-var AssetsPlugin = require('assets-webpack-plugin');
-var assetsPluginInstance = new AssetsPlugin({
-});
-
+const AssetsPlugin = require('assets-webpack-plugin');
+const assetsPluginInstance = new AssetsPlugin({});
+const CleanWebpackPlugin = require('clean-webpack-plugin');
+const CSSSplitWebpackPlugin = require('css-split-webpack-plugin').default;
+const HardSourceWebpackPlugin = require('hard-source-webpack-plugin');
 const { TsConfigPathsPlugin } = require('awesome-typescript-loader');
 
 let paths = require('./webpack-path-base.config')();
@@ -20,18 +21,44 @@ paths.toCopy.map(function (item) {
     item.to = path.join(paths.contentOutput,item.to);
 });
 
-const extractSass = new ExtractTextPlugin({
+const extractCss = new ExtractTextPlugin({
     filename: "css/generated/[name].[contenthash].min.css",
     disable: false,
     allChunks: true
 });
+
+const extractHtml = new ExtractTextPlugin({
+    filename: "html/generated/[name].[contenthash].html",
+    disable: false,
+    allChunks: true
+});
+
+// the path(s) that should be cleaned
+let pathsToClean = [
+    'public/css/app',
+    'public/css/generated',
+    'public/fonts',
+    'public/images',
+    'public/js/app',
+    'public/js/generated',
+    'public/html/generated'
+];
+
+// the clean options to use
+let cleanOptions = {
+    root:     __dirname,
+    exclude:  [],
+    verbose:  true,
+    dry:      false,
+    watch: false		//It could cause issues, e.g. webpack only copies modified images via copy plugin, so we should clear images on watch
+};
+
 
 //https://webpack.js.org/plugins/commons-chunk-plugin/
 //http://stackoverflow.com/questions/39548175/can-someone-explain-webpacks-commonschunkplugin
 
 module.exports = function () {
     return {
-        devtool: 'source-map',  //For debugging purposes
         resolve: {
             extensions: [".ts", ".tsx", ".js", ".json"],
             plugins: [
@@ -50,7 +77,21 @@ module.exports = function () {
             publicPath: paths.public
         },
         plugins: [
-            extractSass,    //Separate css
+            new CleanWebpackPlugin(pathsToClean, cleanOptions),
+            new HardSourceWebpackPlugin(),  //For build cachings. Can cause issues. If so, try disabling it or deleting its cache folder. (default location node_modules/.cache)
+            new webpack.optimize.ModuleConcatenationPlugin(),   //Causes bailouts if array specified as entrypoints
+            extractCss,    //Separate css
+            extractHtml,
+            // new CSSSplitWebpackPlugin({
+            // //     //Splitting also mess up AssetPlugin. Only use, if required
+            //      size: 100, //For IE 9, it is 4095
+            // //     //Make it true if you want to include single css file as before. This single file will have imports for other split files.
+            // //     // Note, this would make additional request and could effect performance for browser which don't even require splitting.
+            // //     // Better way to be to include preserved (non-splitted css) for browser which can work with them, and splitted css for browsers which don't
+            //      imports: 'css/generated/[name]-split.css',
+            //      preserve: true,
+            //      filename: "css/generated/[name]-[part].[ext]"   //Might need to modify regex of OptimizeCssAssetsPlugin
+            // }),
             assetsPluginInstance,
             new CopyWebpackPlugin(paths.toCopy),
             new webpack.optimize.CommonsChunkPlugin({
@@ -70,21 +111,81 @@ module.exports = function () {
             //      Don't put anything else except whats already in entry point. It move any common code which was already part of vendor in vendor chunk and remove it from other chunks
             new webpack.optimize.CommonsChunkPlugin({
                 name: "manifest"
-            }),
-            new webpack.optimize.OccurrenceOrderPlugin()
+            })
         ],
         module: {
             rules: [
                 {
                     test: /\.[s]*css$/,
-                    use: extractSass.extract({
-                        use: [{
-                            loader: "css-loader"
-                        }, {
-                            loader: "sass-loader"
-                        }],
+                    use: extractCss.extract({
+                        use: [
+                            {
+                                loader: "css-loader",
+                                options: {
+                                    importLoaders: 1,
+                                    sourceMap: true
+                                }
+                            }, {
+                                loader: "postcss-loader",
+                                options: {
+                                    sourceMap: true
+                                }
+                            },
+                            {
+                                loader: "sass-loader",
+                                options: {
+                                    sourceMap: true
+                                }
+                            }
+                        ],
                         // use style-loader in development
-                        fallback: "style-loader"
+                        fallback: {
+                            loader: "style-loader",
+                            options: { sourceMap: true }
+                        }
+                    })
+                },
+                {
+                    test: /\.less$/,
+                    use: extractCss.extract({
+                        use: [
+                            {
+                                loader: "css-loader",
+                                options: {
+                                    importLoaders: 1,
+                                    sourceMap: true
+                                }
+                            }, {
+                                loader: "postcss-loader",
+                                options: {
+                                    sourceMap: true
+                                }
+                            },
+                            {
+                                loader: "less-loader",
+                                options: {
+                                    sourceMap: true
+                                }
+                            }
+                        ],
+                        // use style-loader in development
+                        fallback: {
+                            loader: "style-loader",
+                            options: { sourceMap: true }
+                        }
+                    })
+                },
+                {
+                    test: /.pug/,
+                    use: extractHtml.extract({
+                        use: [{
+                            loader: "html-loader"
+                        }, {
+                            loader: "pug-html-loader",
+                            options: {
+                                pretty: true
+                            }
+                        }]
                     })
                 },
                 // All files with a '.ts' or '.tsx' extension will be handled by 'awesome-typescript-loader'.
@@ -118,6 +219,12 @@ module.exports = function () {
              flatpickr: "flatpickr",
              'moment-duration-format': 'moment-duration-format',
              jQuery: "jquery"
-         }
+         },
+        stats: {
+            // Examine all modules
+            maxModules: Infinity,
+            // Display bailout reasons
+            optimizationBailout: true
+        }
     };
 }
