@@ -1,18 +1,19 @@
 const webpackMerge = require('webpack-merge');
-var webpack = require('webpack');
+const webpack = require('webpack');
 const commonConfig = require('./webpack-base.config.js');
-let OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+const OptimizeCssAssetsPlugin = require('optimize-css-assets-webpack-plugin');
 
 module.exports = function() {
     return webpackMerge(commonConfig(), {
+        //devtool: 'source-map',            //Production ready separate sourcemap files with original source code. SourceMaps Can be deployed but make sure to not allow access to public users to them.
+        devtool: 'nosources-source-map',  //Production ready separate sourcemap files with no original source code. SourceMaps Can be deployed securely
         plugins: [
-            new webpack.LoaderOptionsPlugin({
-                minimize: true,
-                debug: false
+            new webpack.LoaderOptionsPlugin({   //can pass shared options to all loaders
+                //passing debug: false, or minimize: true here can cause css minification even without OptimizeCssAssetsPlugin. In that case, it will be handled by css-loader
             }),
             new OptimizeCssAssetsPlugin({
                 //uses processor cssnano by default
-                assetNameRegExp: /\.min.css$/,
+                assetNameRegExp: /\.css$/,
                 cssProcessorOptions: { zindex:false, discardComments: { removeAll: true } }
             }),
             new webpack.DefinePlugin({
@@ -20,7 +21,9 @@ module.exports = function() {
                     'NODE_ENV': JSON.stringify('production')
                 }
             }),
-            new webpack.optimize.UglifyJsPlugin()
+            new webpack.optimize.UglifyJsPlugin({
+                sourceMap: true
+            })
         ]
     })
 }
